requirements.txt
aiogram>=3.13,<4
aiohttp>=3.9
aiosqlite>=0.20
python-dotenv>=1.0
Pillow>=10.0
